requirements.txt
fastapi == 0.64.0
pydantic == 1.7.3
requests == 2.25.1
uvicorn == 0.11.5
jinja2==2.10.3
aiofiles==0.7.0

tortoise-orm==0.17.6

ImageHash == 4.0
Pillow == 6.2.1
loguru == 0.5.3
pandas == 1.2.4
tqdm == 4.45.0